[Feature #161837594] refactor for admin and attendant view

+++ UI/assets/js/getProducts.js
@@ -2,8 +2,6 @@
 
 let table_body = document.querySelector('tbody')
 
-const token = localStorage.getItem('access_token')
-const access_token = "Bearer " + token
 
 if (token === null){
 	window.location.href = "../../index.html"
@@ -11,8 +9,7 @@ if (token === null){
 
 let saveBtn = document.getElementsByClassName('save')[0]
 let cancel = document.getElementsByClassName('cancel')[0]
-	
-//  post the data to via fetchwi
+
 
 window.onload = function allProducts(){
 
@@ -62,33 +59,70 @@ mode: "cors",
 		td_delete.id = data[count]['product_id']
 		saveBtn.id = data[count]['product_id']
 	
-		// console.log(edit[0])	
+		
 		td_detail.addEventListener("click", function (e){
 		e.preventDefault()
 		window.location.href="product_detail.html"
 		localStorage.setItem('product_id', this.id)
 		})
-		td_edit.addEventListener("click", function (e){
-		e.preventDefault()
-		
-		product_id = this.id
-		fetch(`https://store-manger.herokuapp.com/api/v2/products/${product_id}`,{  
-            headers: {
-                'Content-Type': 'application/json',
-                'Access-Control-Allow-Origin':'*',
-                'Access-Control-Request-Method': '*',
-                'Authorization': access_token
-            },
-            method:"GET",
-            mode: "cors",   
-            }).then(function(response){return response.json()})
-            .then(function(data){
-            localStorage.setItem('product', JSON.stringify(data))
-            window.location.href="product_edit.html"
-		})
-           })
+		if (role == "admin"){
+			td_edit.addEventListener("click", function (e){
+				e.preventDefault()
+				
+				product_id = this.id
+				fetch(`https://store-manger.herokuapp.com/api/v2/products/${product_id}`,{  
+					headers: {
+						'Content-Type': 'application/json',
+						'Access-Control-Allow-Origin':'*',
+						'Access-Control-Request-Method': '*',
+						'Authorization': access_token
+					},
+					method:"GET",
+					mode: "cors",   
+					}).then(function(response){return response.json()})
+					.then(function(data){
+					localStorage.setItem('product', JSON.stringify(data))
+					window.location.href="product_edit.html"
+				})
+				   })
+		}else{
+			td_edit.addEventListener('click', function(e){
 
-		td_delete.addEventListener("click", modalCall)
+			e.preventDefault()
+			message = "unauthorized to perform function"
+			let notification = document.getElementById('error-message')
+			notification.innerHTML = `
+			<div Id="error-message-item">
+			<h2>${message}</h2>
+			</div>`;
+			setTimeout(()=> {
+				let message = "";
+				notification.innerHTML = message;
+			}, 3000)
+			})
+		}
+	
+		
+		
+		if (role == "admin"){
+			td_delete.addEventListener("click", modalCall)
+		}else{
+			td_delete.addEventListener("click", function(e){
+			e.preventDefault()
+			message = "unauthorized to perform function"
+			let notification = document.getElementById('error-message')
+			notification.innerHTML = `
+			<div Id="error-message-item">
+			<h2>${message}</h2>
+			</div>`;
+			setTimeout(()=> {
+				let message = "";
+				notification.innerHTML = message;
+			}, 4000)
+			})
+			
+}
+		
 	
 		td_det.appendChild(td_detail)
 		tr.appendChild(td_img)
@@ -120,34 +154,38 @@ function modalCall(e) {
       console.log(product_id)
 
         saveBtn.addEventListener('click', function(e){
-        fetch(`https://store-manger.herokuapp.com/api/v2/products/${product_id}`,{
-            headers: {
-                'Content-Type': 'application/json',
-                'Access-Control-Allow-Origin':'*',
-                'Access-Control-Request-Method': '*',
-                'Authorization': access_token
-            },
-            method:"delete",
-            mode: "cors", 
-            }).then(function(response){return response.json()})
-            .then(function(response){
-				let notification = document.getElementById('error-message')
-				notification.innerHTML = `
-				<div Id="error-message-item">
-				<h2>${response.message}</h2>
-				</div>`
-				;
-				setTimeout(()=> {
-					modal.style.display = "none"
-				}, 2000)
-				setTimeout(()=> {
-					let message = "";
-					notification.innerHTML = message;
-					modal.style.display = "none"
-				}, 4000)
-				window.location.reload()
-
-            })
+		if(role=="admin"){
+			fetch(`https://store-manger.herokuapp.com/api/v2/products/${product_id}`,{
+				headers: {
+					'Content-Type': 'application/json',
+					'Access-Control-Allow-Origin':'*',
+					'Access-Control-Request-Method': '*',
+					'Authorization': access_token
+				},
+				method:"delete",
+				mode: "cors", 
+				}).then(function(response){return response.json()})
+				.then(function(response){
+					let notification = document.getElementById('error-message')
+					notification.innerHTML = `
+					<div Id="error-message-item">
+					<h2>${response.message}</h2>
+					</div>`
+					;
+					setTimeout(()=> {
+						modal.style.display = "none"
+					}, 2000)
+					setTimeout(()=> {
+						let message = "";
+						notification.innerHTML = message;
+						modal.style.display = "none"
+					}, 4000)
+					window.location.reload()
+	
+				})
+		}
+		
+       
 })
        
         cancel.addEventListener('click', function(e){
